lc_requirements.txt
langchain
langchain-core
langchain-google-genai
langgraph
langsmith
streamlit
python-dotenv
pydantic
